Various small stuff
	modified:   ../cvt/gfx/GFXEngineImage.cpp
	modified:   ../cvt/math/Math.cpp
	modified:   ../cvt/math/Math.h
	modified:   ../tests/gfximg/main.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@991 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/math/Math.h
@@ -220,7 +220,7 @@ namespace cvt {
 			return ( ssize_t ) ::random();
 		}
 
-		static inline size_t rand( size_t min, size_t max )
+		static inline int rand( int min, int max )
 		{
 			return min + ( ( max - min ) * ( ( float ) ::random() / ( ( size_t ) RAND_MAX + 1 )  ) );
 		}
